+++ lua/lazyvim/plugins/coding.lua
@@ -82,6 +82,7 @@ return {
         { path = "LazyVim", words = { "LazyVim" } },
         { path = "snacks.nvim", words = { "Snacks" } },
         { path = "lazy.nvim", words = { "LazyVim" } },
+        { path = "nvim-lspconfig", words = { "lspconfig.settings" } },
       },
     },
   },