+++ lua/lazyvim/plugins/extras/lang/python.lua
@@ -38,23 +38,8 @@ return {
               logLevel = "error",
             },
           },
-          keys = {
-            {
-              "<leader>co",
-              LazyVim.lsp.action["source.organizeImports"],
-              desc = "Organize Imports",
-            },
-          },
         },
-        ruff_lsp = {
-          keys = {
-            {
-              "<leader>co",
-              LazyVim.lsp.action["source.organizeImports"],
-              desc = "Organize Imports",
-            },
-          },
-        },
+        ruff_lsp = {},
       },
       setup = {
         [ruff] = function()
@@ -114,6 +99,7 @@ return {
     opts = {
       options = {
         notify_user_on_venv_activation = true,
+        override_notify = false,
       },
     },
     --  Call config for Python files and load the cached venv automatically