+++ lua/lazyvim/plugins/extras/lang/typescript/biome.lua
@@ -1,3 +1,5 @@
+---@module "conform"
+
 ---@diagnostic disable: inject-field
 if lazyvim_docs then
   -- Enable this option to avoid conflicts with Prettier.
@@ -8,6 +10,7 @@ end
 local supported = {
   "astro",
   "css",
+  "scss",
   "graphql",
   -- "html",
   "javascript",
@@ -23,24 +26,35 @@ local supported = {
 }
 
 return {
+  recommended = function()
+    return LazyVim.extras.wants({
+      root = { "biome.json", "biome.jsonc" },
+    })
+  end,
+
   {
-    "mason-org/mason.nvim",
-    opts = { ensure_installed = { "biome" } },
+    "neovim/nvim-lspconfig",
+    opts = {
+      servers = {
+        ---@type lspconfig.settings.biome
+        biome = {},
+      },
+    },
   },
 
   {
     "stevearc/conform.nvim",
     optional = true,
-    ---@param opts ConformOpts
+    ---@param opts conform.setupOpts
     opts = function(_, opts)
       opts.formatters_by_ft = opts.formatters_by_ft or {}
       for _, ft in ipairs(supported) do
         opts.formatters_by_ft[ft] = opts.formatters_by_ft[ft] or {}
-        table.insert(opts.formatters_by_ft[ft], "biome")
+        table.insert(opts.formatters_by_ft[ft], "biome-check")
       end
 
       opts.formatters = opts.formatters or {}
-      opts.formatters.biome = {
+      opts.formatters["biome-check"] = {
         require_cwd = true,
       }
     end,