+++ lua/lazyvim/plugins/extras/ui/mini-indentscope.lua
@@ -55,7 +55,7 @@ return {
     },
   },
 
-  -- disable snacks scroll when mini-indentscope is enabled
+  -- disable snacks scope when mini-indentscope is enabled
   {
     "snacks.nvim",
     opts = {