+++ tests/extras/extra_spec.lua
@@ -5,6 +5,9 @@ local Icons = require("mini.icons")
 
 local Plugin = require("lazy.core.plugin")
 _G.LazyVim = require("lazyvim.util")
+require("lazyvim.config")
+LazyVim.config.get_defaults()
+LazyVim.plugin.setup()
 
 describe("Extra", function()
   local Config = require("lazy.core.config")
@@ -77,10 +80,12 @@ describe("Extra", function()
       it("it has no renamed plugins", function()
         for _, p in pairs(spec.plugins) do
           local short_url = p[1]
-          assert(
-            not LazyVim.plugin.renames[short_url],
-            "Plugin " .. short_url .. " has been renamed to " .. (LazyVim.plugin.renames[short_url] or "")
-          )
+          if short_url then
+            assert(
+              not LazyVim.plugin.renames[short_url],
+              "Plugin " .. short_url .. " has been renamed to " .. (LazyVim.plugin.renames[short_url] or "")
+            )
+          end
         end
       end)
 
@@ -91,8 +96,13 @@ describe("Extra", function()
           local mason = spec.plugins["mason.nvim"]
           local mason_opts = Plugin.values(mason, "opts", false)
 
-          for lsp in pairs(lspconfig_opts.servers or {}) do
+          for lsp, lsp_opts in pairs(lspconfig_opts.servers or {}) do
             local lsp_pkg = lsp_to_pkg[lsp]
+            -- Skip if the LSP server is disabled in the config since mason.nvim won't install it
+            -- and it might still be used for conform.nvim or nvim-lint, etc.
+            if type(lsp_opts) == "table" and lsp_opts.enabled == false then
+              lsp_pkg = false
+            end
             assert(
               not (lsp_pkg and vim.tbl_contains(mason_opts.ensure_installed, lsp_pkg)),
               "LSP server "